Migrate git edit coauthor
